fix(#33): PHP 8.1 update

+++ WebApp/functions_security.php
@@ -35,7 +35,7 @@ class security
                 {return com_create_guid();}
             else
                 {
-                    mt_srand((double)microtime()*10000);
+                    mt_srand((int)microtime()*10000);
                     $charid = strtoupper(md5(uniqid(rand(), true)));
                     $hyphen = chr(45);// "-"
                     $uuid = chr(123)// "{"
